requirements.txt
adblockparser==0.7
bcrypt==3.2.0
beautifulsoup4==4.9.3
cachetools==3.1.0
cairocffi==1.0.2
CairoSVG==2.3.0
cffi==1.12.2
chardet==3.0.4
Click==7.1.2
croniter==0.3.30
cssselect2==0.2.1
dataclasses==0.6
defusedxml==0.5.0
Django==2.2.13
django-cors-headers==2.4.0
django-extensions==2.1.3
django-filter==2.0.0
django-ipware==2.1.0
django-rq==2.0
djangorestframework==3.9.1
djongo==1.2.29
drawSvg==1.1.0
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
httplib2==0.18.0
humanfriendly==4.17
idna==2.8
imagehash==4.2.0
Markdown==2.6.11
netaddr==0.7.19
oauth2client==4.1.3
Pillow==7.1.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
pyjwt==1.7.1
pymongo==3.7.1
psycopg2-binary==2.8.3
python-dateutil==2.8.0
pytz==2018.5
PyYAML==5.1
redis==3.2.1
requests==2.21.0
rq==1.0
rq-scheduler==0.9
rsa==4.0
scapy
selenium==3.141.0
six==1.11.0
sqlparse==0.2.4
tinycss2==1.0.2
ua-parser==0.8.0
uritemplate==3.0.0
urllib3==1.24.2
user-agents==2.0
waitress==1.4.3
webencodings==0.5.1
